requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.2.0
disnake==2.9.0
environs==9.5.0
frozenlist==1.3.3
idna==3.4
marshmallow==3.19.0
multidict==6.0.4
packaging==23.1
python-dotenv==1.0.0
tabulate==0.9.0
yarl==1.9.2
